oidc_handler/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
requests==2.32.2
urllib3==2.2.1
cffi==1.16.0
cryptography==42.0.7
pycparser==2.22
PyJWT==2.8.0
